+++ inst/shiny/modules/tab_nca/setup/settings.R
@@ -460,6 +460,7 @@ settings_server <- function(id, data, adnca_data, settings_override) {
         min_hl_points = input$min_hl_points,
         data_imputation = list(
           impute_c0 = data_imputation$should_impute_c0(),
+          na_imputation_rule = data_imputation$na_imputation_rule(),
           blq_strategy = data_imputation$blq_strategy(),
           blq_imputation_rule = data_imputation$blq_imputation_rule()
         ),